refactor: give panel resizing and hunk navigation named homes
Zero-behavior decomposition of the two components that grew all day:
ChangesPanel's entire resize concern (width formula, storage, drag and
keyboard handlers, aria-bounds observer, separator element) moves to
panel-resize.tsx behind usePanelResize; ReviewDiff's hunk navigation
(currentHunk, goToHunk, the first-hunk landing guard and the deferred
post-commit scroll, with their probe-backed comments) moves to
use-hunk-navigation.ts; the duplicated truncation notice becomes one
TruncationNote. No interface, assertion, or behavior changes.
